refactor(wip): More refactoring of on-chain-data
