+++ src/on-chain-data/airlines/airline.js
@@ -4,14 +4,14 @@ import type {
   TransactionCallbacksInterface,
   TransactionOptionsInterface,
   TxReceiptInterface,
-} from '../interfaces/base-interfaces';
-import type { AirlineInterface, PreparedTransactionMetadataInterface } from '../interfaces/airline-interfaces';
-import Utils from './utils';
-import Contracts from './contracts';
-import RemotelyBackedDataset from './remotely-backed-dataset';
-import StoragePointer from './storage-pointer';
+} from '../../interfaces/base-interfaces';
+import type { AirlineInterface, PreparedTransactionMetadataInterface } from '../../interfaces/airline-interfaces';
+import Utils from '../utils';
+import Contracts from '../contracts';
+import RemotelyBackedDataset from '../remotely-backed-dataset';
+import StoragePointer from '../storage-pointer';
 
-import { InputDataError, SmartContractInstantiationError } from './errors';
+import { InputDataError, SmartContractInstantiationError } from '../errors';
 
 /**
  * Wrapper class for an airline backed by a smart contract on
