+++ src/on-chain-data/hotels/hotel.js
@@ -4,14 +4,14 @@ import type {
   TransactionCallbacksInterface,
   TransactionOptionsInterface,
   TxReceiptInterface,
-} from '../interfaces/base-interfaces';
-import type { HotelInterface, PreparedTransactionMetadataInterface } from '../interfaces/hotel-interfaces';
-import Utils from './utils';
-import Contracts from './contracts';
-import RemotelyBackedDataset from './remotely-backed-dataset';
-import StoragePointer from './storage-pointer';
+} from '../../interfaces/base-interfaces';
+import type { HotelInterface, PreparedTransactionMetadataInterface } from '../../interfaces/hotel-interfaces';
+import Utils from '../utils';
+import Contracts from '../contracts';
+import RemotelyBackedDataset from '../remotely-backed-dataset';
+import StoragePointer from '../storage-pointer';
 
-import { InputDataError, SmartContractInstantiationError } from './errors';
+import { InputDataError, SmartContractInstantiationError } from '../errors';
 
 /**
  * Wrapper class for a hotel backed by a smart contract on
