+++ test/wt-libs/on-chain-data/airlines/wt-airline-index.spec.js
@@ -1,6 +1,6 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
-import WTIndexDataProvider from '../../../../src/on-chain-data/wt-airline-index';
+import WTIndexDataProvider from '../../../../src/on-chain-data/airlines/wt-index';
 import { AirlineDataModel } from '../../../../src/on-chain-data/';
 import testedDataModel from '../../../utils/data-airline-model-definition';
 
@@ -8,7 +8,7 @@ import { WTLibsError } from '../../../../src/errors';
 import { SmartContractInstantiationError, AirlineNotInstantiableError, AirlineNotFoundError,
   RemoteDataReadError, InputDataError } from '../../../../src/on-chain-data/errors';
 
-describe('WTLibs.on-chain-data.WTAirlineIndex', () => {
+describe('WTLibs.on-chain-data.airlines.WTAirlineIndex', () => {
   let dataModel, indexDataProvider;
 
   beforeEach(async function () {
